refactor: moved tests to Asn1DecodedDataRouterTest for test isolation and clarity
